requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.46.0
pypdf==5.0.1
tiktoken==0.7.0
pinecone-client==5.0.1
python-dotenv==1.0.1
